fix: refactor getUserPermissions function
