+++ src/components/CheckForm/CheckForm.tsx
@@ -12,10 +12,10 @@ import { Check, CheckFormValues, CheckType, ROUTES } from 'types';
 import { createNavModel } from 'utils';
 import { AdHocCheckResponse } from 'datasource/responses.types';
 import { useCheckTypeGroupOption } from 'hooks/useCheckTypeGroupOptions';
+import { getUserPermissions } from 'data/permissions';
 import { useCheckTypeOptions } from 'hooks/useCheckTypeOptions';
 import { useCanReadLogs } from 'hooks/useDSPermission';
 import { useLimits } from 'hooks/useLimits';
-import { getUserPermissions } from 'hooks/useUserPermissions';
 import { toFormValues } from 'components/CheckEditor/checkFormTransformations';
 import { CheckJobName } from 'components/CheckEditor/FormComponents/CheckJobName';
 import { ChooseCheckType } from 'components/CheckEditor/FormComponents/ChooseCheckType';
