+++ src/hooks/usePluginPermissions.ts
@@ -1,10 +1,9 @@
 import { hasGlobalPermission } from 'utils';
-
-import { useUserPermissions } from './useUserPermissions';
+import { getUserPermissions } from 'data/permissions';
 
 export function usePluginPermissions() {
   const canEditPlugin = hasGlobalPermission(`plugins:write`);
-  const { canEnablePlugin, canDisablePlugin } = useUserPermissions();
+  const { canEnablePlugin, canDisablePlugin } = getUserPermissions();
 
   return {
     canEnablePlugin: canEditPlugin && canEnablePlugin,
